ZEPPELIN-3206. Move spark integration test to zeppelin-integration
